[#100] Remove Listen trait
